chore: cleanup api entrypoint file, add utils file with helper methods
